アトミックデザインの適用
