+++ src/Authentication/Login.tsx
@@ -7,7 +7,7 @@ import {
   InputAdornment,
   Typography,
 } from '@mui/material';
-import Layout2 from '../layout/Layout2';
+import Layout2 from '../layouts/Layout2';
 import { useState } from 'react';
 import axios from 'axios';
 import Visibility from '@mui/icons-material/Visibility';
