+++ src/main.tsx
@@ -1,16 +1,13 @@
 import ReactDOM from 'react-dom/client';
-import './index.css';
-import App from './App';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Login from './Authentication/Login';
 import Registration from './Authentication/Registration';
 import SendingURL from './Authentication/SendingURL';
-import CreateTask from './Task/CreateTask';
-import ProfileDetail from './Profile/ProfileDetails';
+import ProfileDetail from './pages/ProfileDetails';
 import Complete from './Authentication/Complete';
-import Category from './Category/Category';
-import TaskCategoryList from './Category/TaskCategoryList';
-import TaskList from './Task/TaskList';
+import Category from './pages/Category';
+import TaskCategoryList from './pages/TaskCategoryList';
+import TaskList from './pages/TaskList';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement,
@@ -18,7 +15,6 @@ const root = ReactDOM.createRoot(
 root.render(
   <BrowserRouter>
     <Routes>
-      <Route path="/" element={<App />} />
       <Route path="/login" element={<Login />} />
       <Route path="/registration" element={<Registration />} />
       <Route path="/sendingurl" element={<SendingURL />} />
